extracted phase badge out into its own entity, added to export, updated tests/snapshots
