+++ src/components/PhaseBadge/stories.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { storiesOf } from "@storybook/react";
+
+import PhaseBadge from ".";
+
+storiesOf("PhaseBadge", module).add("PhaseBadge", () => (
+  <PhaseBadge>beta</PhaseBadge>
+));
